Move isFetchingSiteBlocks to state/reader
