+++ client/me/site-blocks/main.jsx
@@ -17,7 +17,7 @@ import MeSidebarNavigation from 'me/sidebar-navigation';
 import PageViewTracker from 'lib/analytics/page-view-tracker';
 import QuerySiteBlocks from 'components/data/query-site-blocks';
 import getBlockedSites from 'state/reader/site-blocks/selectors/get-blocked-sites';
-import isFetchingSiteBlocks from 'state/selectors/is-fetching-site-blocks';
+import isFetchingSiteBlocks from 'state/reader/site-blocks/selectors/is-fetching-site-blocks';
 import getSiteBlocksCurrentPage from 'state/reader/site-blocks/selectors/get-site-blocks-current-page';
 import getSiteBlocksLastPage from 'state/reader/site-blocks/selectors/get-site-blocks-last-page';
 import SiteBlockListItem from './list-item';
