+++ client/state/selectors/test/is-fetching-site-blocks.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import isFetchingSiteBlocks from 'state/selectors/is-fetching-site-blocks';
+import isFetchingSiteBlocks from 'state/reader/site-blocks/selectors/is-fetching-site-blocks';
 
 describe( 'isFetchingSiteBlocks()', () => {
 	test( 'should return true if there is a fetch in progress', () => {
